# Add support for compiling C-extension in parallel.