# References
<!-- Any issues or pull requests relevant to this pull request -->

# Description
<!-- Describe what this request will change/fix and provide any details necessary for reviewers -->
The native `setup.py` support doesn't provide a way to parallelize source compilation, making the compilation time very slow. This PR implements a patch to enable the parallel compilation by learning [grpc](https://github.com/grpc/grpc/blob/master/src/python/grpcio/_parallel_compile_patch.py) and [cpython](https://github.com/python/cpython/blob/31368a4f0e531c19affe2a1becd25fc316bc7501/Lib/distutils/ccompiler.py#L511C9-L511C16). Essentially, this PR implements a method that overloads the default `compile` method to use `multiprocessing` to compile files in parallel.

Local Benchmark:
When building for MacOS in M1 Pro (10 cores), the total compilation time (real-time) using `pip install -e .` is:
Serialized compilation (before this PR): 39.365s
Parallel compilation (after this PR): 11.709s
Improvements: 3.362x

Github Action Benchmark:
Unfortunately, Github Action only provides 2 cores in their virtual environment, so the improvements are not obvious. However, we can still see the improvements by comparing these two actions:
Serialized compilation (before this PR): https://github.com/y-scope/clp-ffi-py/actions/runs/8307187935
Parallel compilation (after this PR): https://github.com/LinZhihao-723/clp-ffi-py/actions/runs/8308605670
We may cherry-pick some builds to check the wheel building time:
cp38-manylinux_aarch64: 386.13s -> 223.95s
cp310-manylinux_i686: 41.50s -> 20.43s
cp311-macosx_arm64: 51.07s -> 22.02s
Note: the time in macos build doesn't have a stable improvement, which might be because the virtualized macos env doesn't have 2 physical cores.

# Validation performed
<!-- What tests and validation you performed on the change -->
Validated the workflow still passed.