@LinZhihao-723 2024-03-17
> How about:
> 
> Add support for compiling C-extension using all cores. (#53)

Shall we use "Add support for compiling C-extension using all cores **in parallel**." or "Add support for compiling C-extension using all cores in **concurrently**."? I think we should be explicit about the parallelism using some key words.