@kirkrodrigues 2024-03-17
Sure, we can use your first suggestion or this one:

Add support for compiling C-extension in parallel. (#53)